RavenDB-24525 Test panel improvements
