+++ src/Raven.Studio/typescript/components/hooks/useResizableWidth.ts
